Support local installation of dependencies
This doesn't change any of the user-visible code, but merely adds an
alternate development process that can be followed, which installs all
dependencies local to the project directory and runs them from there via
`tmuxinator`.

The README was greatly expanded to account for the two different methods
for working on the project, but hopefully not in a way that renders is
cumbersome (I have a tendency to be verbose). To mitigate this, I've
added a "Cookbooks" session which summarizes a few quick commands that
can be run which should hopefully get a user up and running quickly.

Additionally, I added a section to the documentation to expand on how to
contribute; it may not be clear to all people working on the site that
the HTML files must be included in the commits, as normally generated
code is not included, but we need it in this case as we're hosting on a
static page system.