Remove "Sooner is better" from CFP
There is no additional benefit to submitting early; proposals are all
evaluated after the deadline has passed and the CFP is closed, and
earlier submissions are not weighted heavier. In the meantime, it may
discourage people from submitting closer to the deadline, which only
reduces the number of submissions we receive.

So, it's gone now, and the deadline is more clearly specified as a
deadline.