Alter SOC to specifically include gender identity
A potential attendee has pointed out that the existing language is
unclear on the topic of gender assigned at birth vs. gender identity.

This commit, if approved, should provide more clarity. As with any
changes to SOC, this modification should be approved by the organizer
group, as it affects the rules of the conference itself.

The argument for this change is that our conference should be as
inclusive as possible, and attendees made as comfortable as possible,
and confident that we will take any abusive behavior seriously. While
well-intentioned organizers and attendees may automatically assume with
the existing language that gender identity is implied, assumptions do
not translate to clear policy, nor do they provide clear and concise
directions to anyone attempting to enforce the SOC. It is a minor
grammatical change which simply verbally clarifies what the intent
already was, and hopefully ensures more attendees feel welcomed to
attend our conference.