Strengthen the language in the SOC
The language of the SOC is currently passive in terms of describing its
purpose.

This commit, if approved, will make the SOC more active in describing
its role. As with any changes to SOC, this modification should be
approved by the organizer group, as it affects the rules of the
conference itself.

The argument for this change is that the SOC is not just meant to
describe acceptable behaviors, but also to demonstrate that we take
seriously any unacceptable behaviors. It should already be a given that
we expect everyone to behave themselves; this is often the argument
against having an SOC at all. Therefore, being too passive in describing
our standards may lead to the impression that we do not intend to
enforce the rules we've laid out, but are merely providing a list of
common-sense statements about being nice. By being more firm in our
language, we can hopefully guarantee attendees believe we intend for the
rules we've laid out to be followed and will enforce them if necessary.