Update talks list with tutorial schedule
The schedule for the tutorials is complete! Though the tutorials were
already listed in the `talks.csv` file, they were all overlapping with
identical starting times/tracks, and this update fixes that.

It also patches the "block length" for the talks, which was set to 5 and
10, but allowing an extra 30 minutes for general questions per 2.5-hour
block is probably not inaccurate, so they're extended to 6 and 12,
respectively. The speakers don't have to fill this time, of course, but
if they fill it up, attendees won't be left wondering when they heck
they'll be free to get some food. Learning's hungry work!