Break tutorials over lunch
This isn't quite perfect yet; for instance, thursday and Friday have
15-minute offsets, and we may yet want to include mid-morning and
mid-afternoon breaks which are scheduled as well, so attendees are aware
they won't have a marathon. This would probably require adding an
element to the "events.csv" for each break as well, but for now, the
basic layout is there for the tutorials.